+++ meilisearch-http/tests/documents/mod.rs
@@ -1,3 +1,4 @@
 mod add_documents;
 mod delete_documents;
 mod get_documents;
+mod update_documents;
